move test instance to the same package #8
