+++ threading/feature.go
@@ -2,5 +2,5 @@ package threading
 
 type Future interface {
 	isDone() bool
-	get() interface{}
+	Get() interface{}
 }
